Update IntuneBrew.ps1
Added case insensitive comparison of Application name. This would be for cases where someone had previously manually added a software title to Intune and the case differs from the json app feed.

Added code to check for extra permissions granted to the Enterprise application that are not needed for IntuneBrew functionality.

Casting of the $missingPermissions and $extraPermissions as an arrays. Seemed necessary when the app is over provisioned permissions.